requirements.txt
neo4j
requests
python-dotenv
networkx
pandas
matplotlib